Improved: code to use common state module for fetching cycleCount information
