+++ src/store/index.ts
@@ -5,7 +5,6 @@ import mutations from "./mutations";
 import getters from "./getters";
 import actions from "./actions";
 import userModule from "./modules/user";
-import pickerCountModule from "./modules/pickerCount";
 import productModule from "./modules/product";
 import countModule from "./modules/count";
 import { setPermissions } from "@/authorization"
@@ -18,7 +17,7 @@ import { setPermissions } from "@/authorization"
 const state: any = {}
 
 const persistState = createPersistedState({
-  paths: ["user", "pickerCount"],
+  paths: ["user"],
   fetchBeforeUse: true
 })
 
@@ -31,7 +30,6 @@ const store = createStore<RootState>({
   plugins: [ persistState ],
   modules: {
     user: userModule,
-    pickerCount: pickerCountModule,
     product: productModule,
     count: countModule,
   },
