+++ src/store/modules/count/mutation-types.ts
@@ -3,3 +3,5 @@ export const COUNT_LIST_UPDATED = SN_COUNT + "/LIST_UPDATED"
 export const COUNT_QUERY_UPDATED = SN_COUNT + "/QUERY_UPDATED"
 export const COUNT_QUERY_CLEARED = SN_COUNT + "/QUERY_CLEACOUNT_QUERY_CLEARED"
 export const COUNT_STATS_UPDATED = SN_COUNT + "/STATS_UPDATED"
+export const COUNT_UPDATED = SN_COUNT + '/UPDATED'
+export const COUNT_ITEMS_UPDATED = SN_COUNT + '/ITEMS_UPDATED'
