+++ src/store/modules/count/mutations.ts
@@ -21,6 +21,13 @@ const mutations: MutationTree <CountState> = {
       // Parsing the id as in api response we are getting the id in the format `"\"100255\""`
       state.stats[JSON.parse(count.inventoryCountImportId)] = count
     })
+  },
+  [types.COUNT_UPDATED] (state, payload) {
+    state.cycleCounts.list = payload.cycleCount
+    state.cycleCounts.isScrollable = payload.isScrollable;
+  },
+  [types.COUNT_ITEMS_UPDATED] (state, payload) {
+    state.cycleCountItems = payload
   }
 }	
 export default mutations;
