--- src/store/modules/pickerCount/mutation-types.ts
@@ -1,4 +1,0 @@
-export const SN_PICKER_COUNT = "pickerCount"
-export const PICKER_COUNT_UPDATED = SN_PICKER_COUNT + '/COUNT_UPDATED'
-export const PICKER_COUNT_ITEMS_UPDATED = SN_PICKER_COUNT + '/COUNT_ITEMS_UPDATED'
-
