--- src/store/modules/pickerCount/pickerCountState.ts
@@ -1,4 +1,0 @@
-export default interface pickerCountState {
-  cycleCounts: any;
-  cycleCountItems: any;
-}
